% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ons_migration_by_UKborn.R
\name{clean_ons_migration_by_UKborn}
\alias{clean_ons_migration_by_UKborn}
\title{clean_ons_migration_by_UKborn}
\usage{
clean_ons_migration_by_UKborn(
  dir_path = here::here("rawdata"),
  input_file_name = "2.03ltimcountryofbirth2004to2018.xls",
  rtn = TRUE,
  save_to_file = TRUE,
  save_path = here::here("output_data"),
  save_format = "csv"
)
}
\arguments{
\item{dir_path}{input data}

\item{rtn}{TRUE/FALSE}

\item{save_to_file}{TRUE/FALSE}

\item{save_path}{output data}

\item{save_format}{csv or RData}
}
\value{

}
\description{
From ONS migration statistics Excel table
}
